Issue 11731: Added closing of export dialog on form submit(key or submit button click).
git-svn-id: https://svn.imcode.com/imcode/imcms/trunk@23791 bd66a97b-2aff-0310-9095-89ca5cabf5a6

+++ src/main/java/com/imcode/imcms/addon/imagearchive/validator/ExportImageValidator.java
@@ -19,7 +19,7 @@ public class ExportImageValidator implements Validator {
         
         Integer width = command.getWidth();
         if (width != null) {
-            int w = width.intValue();
+            int w = width;
             
             if (w < 1) {
                 width = 1;
@@ -31,7 +31,7 @@ public class ExportImageValidator implements Validator {
         
         Integer height = command.getHeight();
         if (height != null) {
-            int h = height.intValue();
+            int h = height;
             
             if (h < 1) {
                 height = 1;
